mavsdk/requirements.txt
mavsdk==1.3.0
paho-mqtt